initial attempt to break into components
